pyproject.toml
[tool.poetry]
name = "ioi"
version = "0.1.0"
description = "Analysis of the intelligence oversight index survey data"
authors = ["linozen <linus@sehn.tech>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
streamlit = "^0.88.0"
plotly = "^5.3.1"
openpyxl = "^3.0.8"
XlsxWriter = "^3.0.1"
phik = "^0.12.0"

[tool.poetry.dev-dependencies]
jupyterlab-code-formatter = "^1.4.10"
pandas-stubs = "^1.2.0"
isort = "^5.10.1"
black = "^21.10b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
